[#79] fix: PostEntity.price 를 Int64로 변경

+++ ThingLog/Entity/PostEntity+.swift
@@ -11,7 +11,7 @@ import Foundation
 extension PostEntity {
     func update(with post: Post, in context: NSManagedObjectContext) {
         self.title = post.title
-        self.price = Int16(post.price)
+        self.price = Int64(post.price)
         self.purchasePlace = post.purchasePlace
         self.contents = post.contents
         self.giftGiver = post.giftGiver
